move grid search to the libmultilabel module
